put_bytes: casting unsigned int to int
added cast so that we're not comparing an unsigned int with an int

+++ io.c
@@ -22,7 +22,7 @@ void put_bytes(FILE *fp, const void* data, unsigned int len)
 {
   int n = fwrite(data, 1, len, fp);
 
-  if(n != len)
+  if(n != (int) len)
   {
     fprintf(stderr, "Error writing bytes\n");
     exit(EXIT_FAILURE);
